fix - user not shown after deleted

+++ resources/views/users/index.blade.php
@@ -50,10 +50,16 @@
                                 {{--<!--<a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>-->--}}
                                 <a class="btn btn-primary" href="{{ route('users.edit', $user->id) }}">Edit</a>
                                 
-                                {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                                {!! Form::close() !!}
                                 
+                                @if(empty($user->deleted_at))
+                                    {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                                    {!! Form::close() !!}
+                                @else
+                                    <!--{!! Form::open(['method' => 'PATCH','route' => ['users.update', $user->id],'style'=>'display:inline']) !!}
+                                        {!! Form::submit('Restore', ['class' => 'btn btn-success']) !!}
+                                    {!! Form::close() !!}-->
+                                @endif
                             </td>
                         </tr>
                         @endforeach
